+++ cdap-watchdog/src/main/java/co/cask/cdap/metrics/data/TimeValueInterpolatedAggregator.java
@@ -100,7 +100,7 @@ public class TimeValueInterpolatedAggregator implements Iterable<TimeValue> {
 
       // Aggregates values from the same timestamp, using interpolated values if there is not data at the timestamp.
       boolean atEnd = true;
-      int currentTsValue = 0;
+      long currentTsValue = 0;
       Iterator<BiDirectionalPeekingIterator> timeseriesIter = timeseriesList.iterator();
       while (timeseriesIter.hasNext()) {
         BiDirectionalPeekingIterator timeseries = timeseriesIter.next();
